// src/app/features/rappider/components/lib/utils/edit-form/crud-view-form-item-type.enum.ts
export enum CrudViewFormItemType {
  Button = 'button',
  CheckBox = 'checkbox',
  Code = 'code',
  CodeMirror = 'codemirror',
  Date = 'date',
  DateTime = 'datetime',
  DatePicker = 'date-picker',
  DateTimePicker = 'datetime-picker',
  HTMLEditor = 'html-editor',
  Label = 'label',
  List = 'list',
  TextArea = 'textarea',
  TextBox = 'textbox',
  Number = 'number',
  TagBox = 'tag-box',
  Image = 'image',
  Rate = 'rate',
  RowForm = 'row-form',
  Select = 'select',
  TagInput = 'tag-input',
  Slider = 'slider',
  Radio = 'radio',
  RichTextEditor = 'rich-text-editor',
  InputValidatorInfo = 'input-validator-info',
  Object = 'object',
  StringArray = 'string-array',
  ColorPicker = 'color-picker',
  DimensionSelect = 'dimension-select',
  TimePicker = 'time-picker',
  EditFormConfigInput = 'edit-form-config-input',
  JsonArray = 'json-array',
  ListGridConfigInput = 'list-grid-config-input',
  ListGridDataInput = 'list-grid-data-input',
  IconPicker = 'icon-picker',
  RadioGroup = 'radio-group'
}